Feature/server installation (#24)
* Implement automated server installation

* Create async SteamCMD job executor

* Rewrite mod installer service to use SteamCMD job executor

* Implement Workshop mods facade, refactor existing services

* Various fixes, refactoring

+++ src/main/java/cz/forgottenempire/arma3servergui/workshop/services/impl/WorkshopInstallerServiceImpl.java
@@ -2,32 +2,22 @@ package cz.forgottenempire.arma3servergui.workshop.services.impl;
 
 import cz.forgottenempire.arma3servergui.common.Constants;
 import cz.forgottenempire.arma3servergui.common.util.FileSystemUtils;
-import cz.forgottenempire.arma3servergui.common.util.SteamCmdWrapper;
-import cz.forgottenempire.arma3servergui.workshop.entities.DownloadStatus;
-import cz.forgottenempire.arma3servergui.workshop.entities.ErrorStatus;
-import cz.forgottenempire.arma3servergui.system.entities.SteamAuth;
+import cz.forgottenempire.arma3servergui.steamcmd.ErrorStatus;
+import cz.forgottenempire.arma3servergui.steamcmd.entities.SteamCmdJob;
+import cz.forgottenempire.arma3servergui.steamcmd.services.SteamCmdService;
 import cz.forgottenempire.arma3servergui.workshop.entities.WorkshopMod;
 import cz.forgottenempire.arma3servergui.workshop.entities.WorkshopMod.InstallationStatus;
-import cz.forgottenempire.arma3servergui.workshop.repositories.WorkshopModRepository;
-import cz.forgottenempire.arma3servergui.system.services.SteamAuthService;
-import cz.forgottenempire.arma3servergui.workshop.services.WorkshopFileDetailsService;
 import cz.forgottenempire.arma3servergui.workshop.services.WorkshopInstallerService;
-import cz.forgottenempire.steamcmd.SteamCmdParameterBuilder;
-import cz.forgottenempire.steamcmd.SteamCmdParameters;
+import cz.forgottenempire.arma3servergui.workshop.services.WorkshopModsService;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.text.SimpleDateFormat;
+import java.util.Collection;
 import java.util.Date;
-import java.util.HashSet;
 import java.util.Iterator;
-import java.util.Set;
-import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.TimeUnit;
-import java.util.stream.Stream;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -44,57 +34,73 @@ public class WorkshopInstallerServiceImpl implements WorkshopInstallerService {
     @Value("${serverDir}")
     private String serverPath;
 
-    private WorkshopModRepository modRepository;
-    private SteamCmdWrapper steamCmd;
-    private WorkshopFileDetailsService workshopFileDetailsService;
+    private final WorkshopModsService modsService;
+    private final SteamCmdService steamCmdService;
 
-    private SteamAuthService steamAuthService;
-
-    private final ThreadPoolExecutor executor = new ThreadPoolExecutor(1, 1,
-            0L, TimeUnit.MILLISECONDS, new LinkedBlockingQueue<>());
+    @Autowired
+    public WorkshopInstallerServiceImpl(WorkshopModsService modsService, SteamCmdService steamCmdService) {
+        this.modsService = modsService;
+        this.steamCmdService = steamCmdService;
+    }
 
     @Override
-    public void installOrUpdateMod(SteamAuth auth, final WorkshopMod mod) {
-        mod.setInstallationStatus(InstallationStatus.INSTALLATION_QUEUED);
-        modRepository.save(mod);
-
-        executor.submit(() -> {
-            log.info("Starting download of mod {} (id {})", mod.getName(), mod.getId());
-            initializeModDownloadStatus(mod);
-            if (!downloadMod(mod) || !installMod(mod)) {
-                return;
-            }
-            updateModInfo(mod);
-            log.info("Mod {} (id {}) successfully installed ({} left in queue)", mod.getName(), mod.getId(),
-                    executor.getQueue().size());
+    public void installOrUpdateMods(Collection<WorkshopMod> mods) {
+        mods.forEach(mod -> {
+            mod.setInstallationStatus(InstallationStatus.INSTALLATION_IN_PROGRESS);
+            mod.setErrorStatus(null);
+            modsService.saveMod(mod);
         });
+
+        mods.forEach(mod -> steamCmdService.installOrUpdateWorkshopMod(mod)
+                .thenAcceptAsync(steamCmdJob -> handleInstallation(mod, steamCmdJob)));
     }
 
-    private boolean downloadMod(WorkshopMod mod) {
-        DownloadStatus downloadStatus = executeWorkshopModDownload(mod.getId());
-        if (!downloadStatus.isSuccess() || !verifyModDirectoryExists(mod.getId())) {
-            log.error("Failed to download mod {} ({}) ", mod.getName(), mod.getId());
-            mod.setInstallationStatus(InstallationStatus.ERROR);
-            mod.setErrorStatus(downloadStatus.getErrorStatus());
-            modRepository.save(mod);
-            return false;
+    @Override
+    public void uninstallMod(WorkshopMod mod) {
+        File modDirectory = new File(getDownloadPath() + File.separatorChar + mod.getId());
+        try {
+            deleteSymlink(mod);
+            FileUtils.deleteDirectory(modDirectory);
+        } catch (NoSuchFileException ignored) {
+        } catch (IOException e) {
+            log.error("Could not delete mod (directory {}) due to {}", modDirectory, e.toString());
+            throw new RuntimeException(e);
         }
-        return true;
+        log.info("Mod {} ({}) successfully deleted", mod.getName(), mod.getId());
     }
 
-    private boolean installMod(WorkshopMod mod) {
+    private void handleInstallation(WorkshopMod mod, SteamCmdJob steamCmdJob) {
+        if (steamCmdJob.getErrorStatus() != null) {
+            log.error("Download of mod '{}' (id {}) failed, reason: {}",
+                    mod.getName(), mod.getId(), steamCmdJob.getErrorStatus());
+            mod.setInstallationStatus(InstallationStatus.ERROR);
+            mod.setErrorStatus(steamCmdJob.getErrorStatus());
+        } else if (!verifyModDirectoryExists(mod.getId())) {
+            log.error("Could not find downloaded mod directory for mod '{}' (id {}) " +
+                    "even though download finished successfully", mod.getName(), mod.getId());
+            mod.setInstallationStatus(InstallationStatus.ERROR);
+            mod.setErrorStatus(ErrorStatus.GENERIC);
+        } else {
+            log.info("Mod {} (id {}) successfully downloaded, now installing",
+                    mod.getName(), mod.getId());
+            installMod(mod);
+            mod.setInstallationStatus(InstallationStatus.FINISHED);
+        }
+        modsService.saveMod(mod);
+    }
+
+    private void installMod(WorkshopMod mod) {
         try {
             convertModFilesToLowercase(mod);
             copyBiKeys(mod.getId());
             createSymlink(mod);
+            updateModInfo(mod);
+            log.info("Mod '{}' (ID {}) successfully installed", mod.getName(), mod.getId());
         } catch (Exception e) {
-            log.error("Failed to install mod {} ({}) due to: {}", mod.getName(), mod.getId(), e.getMessage());
+            log.error("Failed to install mod {} (ID {})", mod.getName(), mod.getId(), e);
             mod.setInstallationStatus(InstallationStatus.ERROR);
             mod.setErrorStatus(ErrorStatus.IO);
-            modRepository.save(mod);
-            return false;
         }
-        return true;
     }
 
     private void convertModFilesToLowercase(WorkshopMod mod) throws IOException {
@@ -107,11 +113,6 @@ public class WorkshopInstallerServiceImpl implements WorkshopInstallerService {
     private void copyBiKeys(Long modId) throws IOException {
         String[] extensions = new String[]{"bikey"};
         File modDirectory = new File(getModDirectoryPath(modId));
-
-        if (!verifyModDirectoryExists(modId)) {
-            log.error("Can not access mod directory {}", modId);
-            return;
-        }
 
         for (Iterator<File> it = FileUtils.iterateFiles(modDirectory, extensions, true); it.hasNext(); ) {
             File key = it.next();
@@ -136,87 +137,14 @@ public class WorkshopInstallerServiceImpl implements WorkshopInstallerService {
         SimpleDateFormat formatter = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss");
         mod.setLastUpdated(formatter.format(new Date()));
         mod.setFileSize(getActualSizeOfMod(mod.getId()));
-        mod.setInstallationStatus(InstallationStatus.FINISHED);
-        modRepository.save(mod);
-    }
-
-    private void initializeModDownloadStatus(WorkshopMod mod) {
-        mod.setInstallationStatus(InstallationStatus.INSTALLATION_IN_PROGRESS);
-        mod.setErrorStatus(null);
-        modRepository.save(mod);
-    }
-
-    @Override
-    public void uninstallMod(WorkshopMod mod) {
-        File modDirectory = new File(getDownloadPath() + File.separatorChar + mod.getId());
-        try {
-            deleteSymlink(mod);
-            FileUtils.deleteDirectory(modDirectory);
-        } catch (NoSuchFileException ignored) {
-        } catch (IOException e) {
-            log.error("Could not delete mod (directory {}) due to {}", modDirectory, e.toString());
-            throw new RuntimeException(e);
-        }
-        log.info("Mod {} ({}) successfully deleted", mod.getName(), mod.getId());
-    }
-
-    @Override
-    public void updateAllMods(SteamAuth auth) {
-        Set<Long> modIds = getModIds();
-
-        // refresh/update all found mods
-        for (Long modId : modIds) {
-            WorkshopMod mod = modRepository.findById(modId)
-                    .orElseGet(() -> {
-                        // create mods which were not persisted in db
-                        WorkshopMod newMod = new WorkshopMod(modId);
-                        newMod.setName(workshopFileDetailsService.getModName(modId));
-                        return modRepository.save(newMod);
-                    });
-            installOrUpdateMod(auth, mod);
-        }
-    }
-
-    private Set<Long> getModIds() {
-        File modDirectory = new File(getDownloadPath());
-        // find all installed mods in installation folder
-        Set<Long> modIds = new HashSet<>();
-        try (Stream<Path> files = Files.list(modDirectory.toPath())) {
-            files.forEach(p -> {
-                        try {
-                            long id = Long.parseLong(p.getFileName().toString());
-                            modIds.add(id);
-                        } catch (NumberFormatException ignored) {
-                        }
-                    }
-            );
-        } catch (IOException e) {
-            log.error("Failed to collect existing mod directory IDs due to {}", e.toString());
-            throw new RuntimeException(e);
-        }
-
-        // add all missing mods from db
-        modRepository.findAll().stream()
-                .map(WorkshopMod::getId)
-                .forEach(modIds::add);
-        return modIds;
-    }
-
-    private DownloadStatus executeWorkshopModDownload(Long modId) {
-        SteamAuth auth = steamAuthService.getAuthAccount();
-        SteamCmdParameters parameters = new SteamCmdParameterBuilder()
-                .withLogin(auth.getUsername(), auth.getPassword(), auth.getSteamGuardToken())
-                .withInstallDir(downloadPath)
-                .withWorkshopItemInstall(Constants.STEAM_ARMA3_ID, modId, true)
-                .build();
-
-        return steamCmd.execute(parameters);
     }
 
     private void deleteSymlink(WorkshopMod mod) throws IOException {
         Path linkPath = Path.of(getSymlinkTargetPath(mod.getNormalizedName()));
         log.info("Deleting symlink {}", linkPath);
-        Files.delete(linkPath);
+        if (Files.isSymbolicLink(linkPath)) {
+            Files.delete(linkPath);
+        }
     }
 
     private boolean verifyModDirectoryExists(Long modId) {
@@ -242,25 +170,5 @@ public class WorkshopInstallerServiceImpl implements WorkshopInstallerService {
                 + File.separatorChar + "workshop"
                 + File.separatorChar + "content"
                 + File.separatorChar + Constants.STEAM_ARMA3_ID;
-    }
-
-    @Autowired
-    public void setSteamCmd(SteamCmdWrapper steamCmd) {
-        this.steamCmd = steamCmd;
-    }
-
-    @Autowired
-    public void setModRepository(WorkshopModRepository modRepository) {
-        this.modRepository = modRepository;
-    }
-
-    @Autowired
-    public void setWorkshopFileDetailsService(WorkshopFileDetailsService workshopFileDetailsService) {
-        this.workshopFileDetailsService = workshopFileDetailsService;
-    }
-
-    @Autowired
-    public void setSteamAuthService(SteamAuthService steamAuthService) {
-        this.steamAuthService = steamAuthService;
     }
 }
